trivial fixes from building spamc with 'gcc  -Wall -Wextra -pedantic' -- fix C++-style comments to be C90-compliant, and use explicit casts when compared signed and unsigned values
git-svn-id: https://svn.apache.org/repos/asf/spamassassin/trunk@425009 13f79535-47bb-0310-9956-ffa450edef68

+++ spamc/libspamc.c
@@ -612,16 +612,16 @@ static int _message_read_bsmtp(int fd, struct message *m)
     for (i = j = 0; i < (unsigned int) m->msg_len; i++) {
 	if (prev == '\n' && m->msg[i] == '.') {
 	    /* Dot at the beginning of a line */
-            if (((i+1) == m->msg_len)
-                || ((i+1) < m->msg_len && m->msg[i + 1] == '\n')
-                || ((i+2) < m->msg_len && m->msg[i + 1] == '\r' && m->msg[i + 2] == '\n')) {
+            if (((int) (i+1) == m->msg_len)
+                || ((int) (i+1) < m->msg_len && m->msg[i + 1] == '\n')
+                || ((int) (i+2) < m->msg_len && m->msg[i + 1] == '\r' && m->msg[i + 2] == '\n')) {
 		/* Lone dot! That's all, folks */
 		m->post = m->msg + i;
 		m->post_len = m->msg_len - i;
 		m->msg_len = j;
 		break;
 	    }
-	    else if ((i+1) < m->msg_len && m->msg[i + 1] == '.') {
+	    else if ((int) (i+1) < m->msg_len && m->msg[i + 1] == '.') {
 		/* Escaping dot, eliminate. */
 		prev = '.';
 		continue;
